Unify all shared code under the same `:taskodoro-core` module
The original idea was to name the shared module as `:taskodoro` but we
cannot due to this issue in Gradle:
    https://github.com/gradle/gradle/issues/16608

In any case, the package names and iOS framework are named correctly.
