+++ apps/ios/App/App/Assets.xcassets/Contents.json
