+++ apps/ios/Taskodoro.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:Taskodoro/Taskodoro.xcodeproj">
+      location = "group:App/App.xcodeproj">
    </FileRef>
 </Workspace>
