+++ settings.gradle.kts
@@ -26,8 +26,7 @@ rootProject.name = "Taskodoro"
 
 include(":apps:android:app")
 
-include(":tasks")
-include(":storage")
+include(":taskodoro-core")
 
 enableFeaturePreview("VERSION_CATALOGS")
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
