+++ taskodoro-core/src/androidMain/AndroidManifest.xml
