+++ taskodoro-core/src/commonMain/kotlin/com/taskodoro/tasks/model/TaskValidationResult.kt
